Some helpers on Grid
